Factor out gitlab configuration from command into reusable base class.
